+++ builtInC.h
@@ -9,21 +9,21 @@ size_t numOfAllTheSlashesInPath(const char* currentPath);
 int findIndexOfLastSlash(const char* currentPath);
 char* returnUpPath(const char* currentPath);
 void update_PWD_OLDPWD(const char* newPWD, const char* newOLDPWD);
-int cdCommand(char** argv ,size_t indexArgv, size_t argc);
+int cdCommand(char** argv ,size_t indexArgv, int argc);
 //-----------------------------CD_COMMAND---------------------------------
 
 
 
 //-----------------------------EXEC_COMMAND---------------------------------
 void run_execvp(char** argv,size_t indexArgv, char** newArgv);
-void execCommand(char** argv ,size_t indexArgv, size_t argc);
+void execCommand(char** argv ,size_t indexArgv, int argc);
 //-----------------------------EXEC_COMMAND---------------------------------
 
 
 //-----------------------------PWD_COMMAND---------------------------------
 
 //argc must not exceed 1
-void pwdCommand(size_t argc);
+void pwdCommand(int argc);
 //-----------------------------PWD_COMMAND---------------------------------
 
 
@@ -31,7 +31,7 @@ void pwdCommand(size_t argc);
 //-----------------------------EXIT_COMMAND---------------------------------
 
 //argc must not exceed 1
-void exitCommand(size_t argc);
+void exitCommand(int argc);
 //-----------------------------EXIT_COMMAND---------------------------------
 
 #endif
